feat(edition-folio): add basic support for folio visualisation
# Why is this change needed?
Prior to this change, it was not clear, where the sheets of a sketch are
located on a manuscript. The user had no orientation about the overall
structure of the files.

# How does it address the issue?
This change provides a visualisation feature for the manuscript folios,
that uses a json-file input and snapsvg to render an svg overview.

# Provide links to any relevant tickets, articles or other resources